backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.3
google-generativeai==0.7.2
python-dotenv==1.0.1
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
Werkzeug==3.0.3
dnspython==2.6.1
networkx==3.3
firebase-admin==6.5.0

